+++ demisto_sdk/commands/common/errors.py
@@ -909,10 +909,10 @@ class Errors:
     @staticmethod
     @error_code_decorator
     def integration_is_deprecated_and_used(integration_name: str, commands_dict: dict):
-        erorr_str = f"{integration_name} integration contain deprecated commands that are being used:\n"
+        erorr_str = f"{integration_name} integration contains deprecated commands that are being used by other entites:\n"
         for command_name, command_usage_list in commands_dict.items():
-            current_command = '\n'.join(command_usage_list)
-            erorr_str += f"{command_name} is being used in the following locations:\n{current_command}\n"
+            current_command_usage = '\n'.join(command_usage_list)
+            erorr_str += f"{command_name} is being used in the following locations:\n{current_command_usage}\n"
         return erorr_str
 
     @staticmethod
@@ -2271,8 +2271,8 @@ class Errors:
     @staticmethod
     @error_code_decorator
     def playbook_is_deprecated_and_used(playbook_name: str, files_list: list):
-        files_list = '\n'.join(files_list)
-        return f"{playbook_name} playbook is deprecated and being used in the following files:\n{files_list}"
+        files_list_str = '\n'.join(files_list)
+        return f"{playbook_name} playbook is deprecated and being used by the following entites:\n{files_list_str}"
 
     @staticmethod
     @error_code_decorator
@@ -2309,8 +2309,8 @@ class Errors:
     @staticmethod
     @error_code_decorator
     def script_is_deprecated_and_used(script_name: str, files_list: list):
-        files_list = '\n'.join(files_list)
-        return f"{script_name} script is deprecated and being used in the following files:\n{files_list}"
+        files_list_str = '\n'.join(files_list)
+        return f"{script_name} script is deprecated and being used by the following entites:\n{files_list_str}"
 
     @staticmethod
     @error_code_decorator